imou-telegram-alert-mvp/requirements.txt
opencv-python==4.10.0.84
ultralytics==8.3.56
python-dotenv==1.0.1
requests==2.32.3
numpy==2.1.1
